requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.34
asyncpg==0.29.0
aiosqlite==0.20.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
email-validator==2.2.0
pytest==8.3.2
httpx==0.27.2
python-dotenv==1.0.1
